Highlight inside of mathjax object.
Added Highlight object to the Hub.  This contains function
getHighlighter that gets the object to use for doing highlights.  It is
specific to the jax object so this is passed in on the call.

The highlighter itself contains calls for:
setHighlightStyle to specify styles to use for different highlight
types;
highlightNodes to highlight a specific type for given range;
clearHighlights to clear all highlights;
clearHighlightNodes to clear specific highlihts;
setClearLastHighlight to set whether a new highlight should clear the
old one.